Session 16.9: float -> integer microUSDC end-to-end
Money flows are now integer microUSDC across DB, Python internals, JWT
claims, and the wire format. Float USDC contained to one well-marked UI
spot (WalletChip pending-amount delta) and config parsing boundaries.
Every `+ 1e-9` epsilon and tolerance comparison in budget/spent paths
is gone — comparisons are bit-exact integer equality.

Backend:
- New services/money.py with to_micro() + micro_str() helpers.
- models.py: Numeric(18,6) + Mapped[float] -> BigInteger + Mapped[int]
  on Campaign.cpm_price/budget/spent/protocol_fee_amount and
  Settlement.amount_usdc. Field names kept (cpm_price stays as
  microUSDC per 1000 plays — production needs CPM as a pricing unit
  separate from per-play cost).
- services/calc.py: compute_quote() returns Quote with _micro suffixes.
  Protocol fee uses basis points (250 bps = 2.5%) for exact int math.
- services/tokens.py: ProofContextClaims.amount_usdc -> amount_micro.
  Added v=2 schema field; v=1 tokens fail to decode (TTL covers deploy).
- routers/proof.py: atomic UPDATE budget guard is exact-int
  (`budget - spent >= :amount_micro`); COMPLETED-flip is bit-precise.
- services/x402.py: build_payment_requirements(amount_micro: int).
  Drops the int(round(* 1e6)) conversion.
- services/solana.py: get_usdc_balance -> get_usdc_balance_micro
  (returns int via value.amount, not ui_amount). build_usdc_transfer_tx
  takes amount_micro: int directly.
- schemas.py: every money field is MicroStr (= str). Pydantic
  serializes integer micro as "422000"-style strings.

Frontend:
- New lib/money.ts with BigInt-native formatUsdc / sumMicro / subMicro
  / cmpMicro. parseUsdc reserved for the WalletChip UI delta only.
- TS types in lib/aggregations.ts flipped number -> string for money.
- ~15 .toFixed display sites updated across Overview, Campaigns,
  CampaignCard, WalletChip, wizard/StepCalculator, wizard/StepReview.
- StepReview drops the *1.05 slack on x402 client amount — exact int
  micro from server quote eliminates the float-drift safety margin.

Scripts:
- audit_ledger.py + audit_ledger_verbose.py rewritten for exact `==`
  comparison instead of < tolerance.
- e2e_demo.py: _seed_campaign(budget_micro, cpm_micro), assertions on
  int micro, no epsilon.
- sweep_to_treasury, sweep_helpers, cleanup_drift{,_reverse},
  probe_sponsorship, topup_campaigns, check_balance: all migrated.

Validated:
- e2e_demo.py: 25/25 steps pass (multi-campaign batch isolation,
  refund-with-pending, drained-budget auto-completion).
- audit_ledger.py: zero DRIFT, zero SHORT, zero IN-FLIGHT.
- Multi-tenant soak: 1671 plays across 6 campaigns under 2 advertisers,
  every campaign matches expected to the micro (zero diff).
- tsc --noEmit clean on frontend.
- grep "1e-9|epsilon|tolerance" backend/app: 3 hits, all in comments
  documenting the historical fix; no live math.

Closes PLAN.md Must-fix #3. Postgres migration in Session 17 maps
BigInteger -> BIGINT directly with no float-precision shims.

+++ backend/app/models.py
@@ -1,7 +1,7 @@
 from datetime import date, datetime, timezone
 from enum import Enum
 
-from sqlalchemy import JSON, Date, DateTime, ForeignKey, Integer, Numeric, String
+from sqlalchemy import JSON, BigInteger, Date, DateTime, ForeignKey, Integer, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from .database import Base
@@ -45,9 +45,12 @@ class Campaign(Base):
     name: Mapped[str] = mapped_column(String)
     creative_url: Mapped[str] = mapped_column(String)
     creative_id: Mapped[str] = mapped_column(String)
-    cpm_price: Mapped[float] = mapped_column(Numeric(18, 6))
-    budget: Mapped[float] = mapped_column(Numeric(18, 6))
-    spent: Mapped[float] = mapped_column(Numeric(18, 6), default=0)
+    # Session 16.9: money is stored as integer microUSDC (1 USDC = 1e6 micro).
+    # cpm_price is "microUSDC per 1000 plays" — e.g. $0.50 CPM = 500_000.
+    # Per-play cost is derived: cpm_price // 1000.
+    cpm_price: Mapped[int] = mapped_column(BigInteger)
+    budget: Mapped[int] = mapped_column(BigInteger)
+    spent: Mapped[int] = mapped_column(BigInteger, default=0)
     status: Mapped[str] = mapped_column(String, default=CampaignStatus.DRAFT.value, index=True)
 
     wallet_id: Mapped[str] = mapped_column(String)
@@ -67,8 +70,8 @@ class Campaign(Base):
     # x402 funding tx; transferred from the campaign wallet to a dedicated
     # PROTOCOL_REVENUE_WALLET right after settle confirms. Non-refundable —
     # refund only returns budget - spent.
-    protocol_fee_amount: Mapped[float | None] = mapped_column(
-        Numeric(18, 6), nullable=True
+    protocol_fee_amount: Mapped[int | None] = mapped_column(
+        BigInteger, nullable=True
     )
     protocol_fee_tx_hash: Mapped[str | None] = mapped_column(String, nullable=True)
 
@@ -84,7 +87,7 @@ class Settlement(Base):
     campaign_id: Mapped[str] = mapped_column(ForeignKey("campaigns.id"), index=True)
     nonce: Mapped[str] = mapped_column(String, unique=True, index=True)
     publisher_wallet: Mapped[str] = mapped_column(String)
-    amount_usdc: Mapped[float] = mapped_column(Numeric(18, 6))
+    amount_usdc: Mapped[int] = mapped_column(BigInteger)
     tx_hash: Mapped[str | None] = mapped_column(String, nullable=True)
     status: Mapped[str] = mapped_column(String, default=SettlementStatus.CONFIRMED.value)
     # device_id captures which screen the bid was issued for (from
